Large refactor to SDL rendering.
Everything is now done in terms of a `Screen`, which is conceptually a
grid of cells, with `ScreenPos` and `ScreenRect` representing regions in
this grid. Pixel coordinates are only used in a couple of methods on
`Screen` which do the actual rendering.

This brings the UI down from using three coordinate systems (`Point`,
`ScreenPos`, and pixels) to effectively two (just `Point` and
`ScreenPos`).