More pleasing text rendering.
The font image has 12x12 tiles, however almost all are 8x12, including
all the alphanums. This leads to a huge amount of letter spacing, which
is hard to read. As the only tiles which are 12 wide are non-letters,
it's safe to just truncate the tile to the central 8 pixels.

This commit also allows vertical truncation, although that is currently
unused.